Game rebuilds ok for deploy
Rebuilding a game in deploy is working. Started to fail because we have
the terrain on the board that is set randomly. Now the writer also
outputs that and the correct value is set.

Also, on deploy, we don't have turn actions. So I made
str->raw-turn-actions ready to receive nil.