Simple game now completely reloaded
Had to add `first-player` to the game and persist it. I Though that I
could infer the first player. but for that I would need to rely heavily
on action format. Now when the game is started, that key is set.

Also added to `game/start-battle` an overload that gets the first
player. This is mainly to configure test scenarios and for the reader to
be able to start the game.

We're still missing some edge cases:
* What if no player deploys?
* What if only one player deploys?