Properly load games in deploy state
So, I didn't add the stash to the game properties, because I could get
that information from the deploy actions. However, if one or the two
players don't deploy, I don't have the actions to build the stash.

So the writer now outputs the stash, if the player did not deploy yet.
And the reader tries to get the stash from the attributes or from the
deploy actions.

Also add several tests for edge cases.