Add test.check scenario for writer/reader
With the base tests ready, now was time for a test.check spec with
proper random stash generation. And with that stash, write and read the
game and see if it's the same.

Had a hard time doing this, mainly because when a turn is completed, the
result was being stored on a vector on the game. This result had the
game itself... and all this info was creating circular references on the
game map. Meaning the outputing the game was never ending, or even just
a compare was never ending.

The board was there at start so we could see the board generated at
every action. But having a deterministic turn algo, makes it easy to
just replay the actions, and we don't need to store them.

Or changes were just for compatibility. The engine supports for example
using :rain or "rain", and I had some generators outputing "rain", but
the reader would force :rain. Now all handle :rain.