Use pmap on checking option value
We generate several options per element, and then we play firingsquad
against each option to improve the value. This map is now parallel, and
in my machine with criterium, the different went from 35ms median to
25ms.

Add host/parallel-map because on JS we don't have pmap. This way this fn
will try to run the map in parallel if possible, if not will use the
regular map.