Make alamo more afraid to loose units
Make alamo run away if in danger of loosing units or if an attack will
be damaging (like going against strikeback).

This meant that we needed do increase the number of options per element
to process to 10. And also fix a bug. The move-options generator was
getting the amount to generate, but instead of using it was just
returning 1 option. This meant that not all escape options were
available for processing.

Overall, on scenarios with few units and where there is no super range
unit (like kamikaze), alamo is behaving pretty well.