Refactor game creation to receive different stashes for each player
This commit introduced breaking changes, hence the bump in the major
version.

The main motivation for this refactor is that it seems much more likely
that a game needs to be created with different stashes for each player.
This flaw was revealed by the lack of flexibility while writing tests to
support different game modes, in which it is important to create games
in different conditions for each player and that was not possible.

The downside of this change is that it made the creation of a game slightly
less convenient.

Issue #9