Support game draws in annihilation mode
Adding the draw functionality to the annihilation mode led to the refactoring
of the win conditions for this mode. This refactoring might seem somewhat
strange at first glance, since it makes the function definition generic for
more than two players, but that was not the goal. It was just a side
product of trying to find a more functional definition for the win and
drawing conditions.

Issue #9