Make it possible to update the stash of a player in a given game
Motivation:
In the upcoming supernova mode, it should be possible to add units to a
player's stash during game creation. This function was built to be more
generic than the simple addition of units to the stash. The rationale
behind this is explained in the Rationale section below.

Modifications:
Add `update-stash` function to the `game` namespace.

Rationale:
In order to fulfil the requirements needed for the supernova game mode,
one could've simply added a `game/add-units` function, wrapping a call to
`stash/add-units`.
The decision to do it like this comes from the fact that that other approach
is not very composable. Any time that a new requirement comes up and a new
operation needs to be done on one of the game's field, a new function would
need to be added to the `game` namespace and to the related namespace (`stash`
in this case). This design indicates poor use of function composition
and would make the `game` namespace too bloated in no time.
Instead of doing this, it was decided that a single modifying function
would be provided for the game stash (`update-stash`) and it is up for
the client to compose it with any other function that modifies a stash
(like `stash/add-units`).

Issue #9