Add star units to each player's stash on a new supernova game
Motivation:
Being able to create new supernova games and provide an easy way
to add mode specific behavior on game creation.

Modifications:
* Fix minor syntactic issues and typos;
* Add hook to game creation;
* Add star unit metadata and add it to the list of available units;
* Use new game creation hook to add the star units to each player's
  stash on supernova mode;
* Solve circular dependency created between the game and game-mode
  namespaces, by moving game creation into the game-progress namespace;
* Rename the game/create function to game-progress/new-game.

Remarks:
* To solve the circular dependency issue, a few strategies were tried but
  they all led to a similar circular dependency. The game namespace is heavily
  depended upon and the only solution was to not have it depend on any other
  game related namespace (like game-mode).
  The move to the game progress namespace seemed natural since a newly
  created game can be seen as the very first stage of the game.

Further improvement:
When trying to figure out the solution for the circular dependency
issue and thinking about the right place to put the create function, it
became clear that the functions that should be publicly exposed by obb-rules
are not well-defined.
A client of this library would only need to know a small subset of the
functions and it would be great to make an effort and think of a proper
interface, especially considering that obb-rules will be the core
component of the game.

Issue #9