Refactor history to have extra info like the player
Before the history would only have a coll of actions for each turn. This
refactor changed that to a coll of history items. A history item is a
map with `:actions` (the previous actions), and now was added a
`:player` with the player that performed the actions.

In the future we can store more information here, for example, the
amount of time that the player took to play.

Also update the writer/reader to expect the player key as the first
thing on a turn row.