Allow reload of a custom text game on the player
Before, the current game was being written as text on the game, on each
turn. Now, I replaced the raw text with an editable textarea and added a
load button. If the textarea is changed, that new str is stored, and
when the button is pressed the current game is replaced with the loaded
one.

Still has some issues (sometimes the Ai doesn't play on a loaded game),
but it's already a great way to have a visual representation of any
game.