Improve type hints
Replace Integer for primitive long and add more type hints to the unit.

A long is a primitive, and with such, the jvm can do primitive very fast
math. With the Integer reference type, it has to box/unbox it. If it has
no type hint, clojure will use heavy reflection to find out the type.

More details about performance issues:
https://www.youtube.com/watch?v=iQwQXVM6oiY