Store demo game-options instead of game mode
I was storing on the webapp state the mode for the ai vs ai game.
Whenever a game started, I'd use that mode to start a new game.

Changed that logic to store an actual map with the full game options.
By default it uses `game-progress/default-new-game-options`. For the
supernova mode, it just associates the :supernova mode to the defaults.