Require that game mode is specified on game creation
Context:
While trying to add the game mode to `obb-rules.serializer.reader`, I
found that the reader was working even when the game mode was not
part of the input string. The reason was that after the string being
parsed, `game-progress/new-game` was being called and this function
specified a default value for the game mode, causing the game to be
loaded without errors.
In my understanding, the reader should only accept input strings that
contain a complete description of the game, including the game mode.
After giving this some thought, I decided that having a default mode
for the game does not make much sense. I recognise that Annihilation
will probably be the most common mode, but since the rules are so
different between modes, I think it makes sense to require that the
client always provides it (even for random games).
This also prevents the problem we had with the reader, since the
partial definition of the game would no longer work.

Motivation:
Adding support to read the game mode from the string representation
generated by `obb-rules.serializer.writer`.

Modifications:
* Remove default value for the mode on game creation;
* Make the mode required for randomly generated games;
* Fix all tests to respect new game creation argument requirements.
* Update documentation examples

Issue #34